+++ src/comp6521/lab/com/Records/RecordElement.java
@@ -13,8 +13,8 @@ public abstract class RecordElement {
 	public void   setString(String data) { assert(false);                    }
 	public int    getInt()	             { assert(false); return 0;          }
 	public void   setInt(int val)        { assert(false);                    }
-	public float  getFloat()             { assert(false); return 0;          }
-	public void   setFloat(float val)    { assert(false);                    }
+	public double getFloat()             { assert(false); return 0;          }
+	public void   setFloat(double val)   { assert(false);                    }
 	public Date   getDate()              { assert(false); return new Date(); }
 	public void   setDate(Date val)      { assert(false);                    }
